Adding Q-Value Algorithm
Note: The algorithm is partially complete. It runs based of my defined
Reward Function, so the support for the Episode Analysis doesn't exist.
I will try to fix that this weekend. It is tuned to run 50 episodes and
return the number of steps it takes to reach the goal state. More info
in the email.