Cleaned up ObjectClass data member names (removed trailing '_').
Also added a new general purpose grid world domain that can have an
arbitrarily set map and can easily changed to use stochastic transition
dynamics. This will be useful as a "common" domain for the distributed
package. It also can effectively replace the FourRooms domain (to that
end, I added a special method that will set the map to the classic 4
rooms map).